Move __tests__ inside src
